SDI-360 Update Nomis from incentive creation domain event

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/prisonertonomisupdate/services/NomisApiService.kt
@@ -91,5 +91,5 @@ data class CreateIncentiveDto(
 
 data class CreateIncentiveResponseDto(
   val nomisBookingId: Long,
-  val nomisIncentiveSequence: Long,
+  val nomisIncentiveSequence: Int,
 )
